+++ wazo_call_logd/database/queries/recording.py
@@ -5,6 +5,10 @@ from .base import BaseDAO
 from ..models import Recording
 
 
+class _UselessQuery(Exception):
+    pass
+
+
 class RecordingDAO(BaseDAO):
     def create_all(self, recordings):
         with self.new_session() as session:
@@ -27,20 +31,27 @@ class RecordingDAO(BaseDAO):
             query.delete(synchronize_session='fetch')
             session.flush()
 
+    def delete_media_by(self, **kwargs):
+        with self.new_session() as session:
+            query = session.query(Recording)
+            try:
+                query = self._apply_filters(query, kwargs)
+            except _UselessQuery:
+                return
+
+            recordings = query.all()
+            for recording in recordings:
+                recording.path = None
+                session.flush()
+                session.expunge(recording)
+
     def find_all_by(self, **kwargs):
         with self.new_session() as session:
             query = session.query(Recording)
-
-            if 'call_log_ids' in kwargs:
-                if not kwargs['call_log_ids']:
-                    return []
-                query = query.filter(Recording.call_log_id.in_(kwargs['call_log_ids']))
-
-            if 'call_log_id' in kwargs:
-                query = query.filter(Recording.call_log_id == kwargs['call_log_id'])
-
-            if 'uuid' in kwargs:
-                query = query.filter(Recording.uuid == kwargs['uuid'])
+            try:
+                query = self._apply_filters(query, kwargs)
+            except _UselessQuery:
+                return []
 
             recordings = query.all()
             for recording in recordings:
@@ -50,20 +61,27 @@ class RecordingDAO(BaseDAO):
     def find_by(self, **kwargs):
         with self.new_session() as session:
             query = session.query(Recording)
-
-            if 'call_log_ids' in kwargs:
-                if not kwargs['call_log_ids']:
-                    return
-                query = query.filter(Recording.call_log_id.in_(kwargs['call_log_ids']))
-
-            if 'call_log_id' in kwargs:
-                query = query.filter(Recording.call_log_id == kwargs['call_log_id'])
-
-            if 'uuid' in kwargs:
-                query = query.filter(Recording.uuid == kwargs['uuid'])
+            try:
+                query = self._apply_filters(query, kwargs)
+            except _UselessQuery:
+                return
 
             recording = query.first()
             if not recording:
                 return
             session.expunge(recording)
             return recording
+
+    def _apply_filters(self, query, params):
+        if 'call_log_ids' in params:
+            if not params['call_log_ids']:
+                raise _UselessQuery()
+            query = query.filter(Recording.call_log_id.in_(params['call_log_ids']))
+
+        if 'call_log_id' in params:
+            query = query.filter(Recording.call_log_id == params['call_log_id'])
+
+        if 'uuid' in params:
+            query = query.filter(Recording.uuid == params['uuid'])
+
+        return query